Add general cleaning functions to transform.py
This commit adds a set of general cleaning functions to the transform.py file.
These functions can be used to clean data in a variety of formats and are
designed to be flexible and reusable. The functions include:
- get_raw_data
- clean_data
- remove_period
- clean_station_id
- cast_values_to_str
- change_column_name
- remove_null_values

These functions will help to improve the overall quality and consistency of
the data being processed.