Implement preprocessing functions for  data transformation of raw data
This commit implements a set of preprocessing functions to facilitate efficient
transformation of raw data. These functions are designed to clean, format and
structure the raw data, making it ready for further analysis and processing.
Creates the the final data used for analysis in the processed directory
in the data folder.